server connected to db
