+++ server/graphql/resolvers/index.js
@@ -1,4 +1,6 @@
-const { v4: uuidv4}  = require('uuid');
+const { v4: uuidv4 } = require('uuid');
+
+
 
 const resolvers = {
   Query: {
@@ -8,8 +10,10 @@ const resolvers = {
     user: (parent, { id }, { models }) => {
       return models.users[id]
     },
-    users: (parent, args, { models }) => {
-      return Object.values(models.users);
+    users: async (_, __, { models }) => {
+      console.log(models)
+      const users = await models.users.find()
+      return users
     },
     item: (parent, { id }, { models }) => {
       return models.items[id]
@@ -28,8 +32,23 @@ const resolvers = {
       return Object.values(models.items).filter(item => item.userID === user.id)
     }
   },
+
+
+  // async createCatcall(_, { catcall }) {
+  //   const { type, geometry, properties} = catcall;
+  //   const createdCatcall = await Catcall.create({ type, geometry, properties });
+  //   return createdCatcall;
+  // },
+
+
   Mutation: {
-    createItem: (parent, { name }, {me, models}) => {
+
+    createUser: async (_, { user }, { models }) => {
+      const createdUser = await models.users.create({ id: 1, username: user });
+      return createdUser;
+    },
+
+    createItem: (parent, { name }, { me, models }) => {
       let id = uuidv4();
       console.log(id);
       const item = {
@@ -44,12 +63,12 @@ const resolvers = {
     },
     deleteItem: (parent, { id }) => {
       let { [id]: item, ...otherItems } = items;
-      if(!item) {
+      if (!item) {
         return false;
-      } 
+      }
       model.items = otherItems;
       return true;
-      
+
     }
   }
 };
