+++ server/graphql/schemas/mutations.schema.js
@@ -1,5 +1,6 @@
 module.exports = `
 type Mutation {
+    createUser(user: String!): User!
     createItem(name: String!): Item!
     deleteItem(id: ID!): Boolean!
   }`;
