+++ server/models/index.js
@@ -1,30 +1,33 @@
-let users = {
-  1: {
-    id: '1',
-    username: 'Robin Wieruch',
-    itemIds: [1],
-  },
-  2: {
-    id: '2',
-    username: 'Dave Davids',
-    itemIds: [2],
-  },
-};
- 
-let items = {
-  1: {
-    id: '1',
-    name: 'Hello World',
-    userId: '1',
-  },
-  2: {
-    id: '2',
-    name: 'By World',
-    userId: '2',
-  },
-};
- 
-module.exports =  {
-  users,
-  items
-};
+'use strict';
+
+const fs = require('fs');
+const path = require('path');
+const Sequelize = require('sequelize');
+const basename = path.basename(__filename);
+const db = {};
+const dbconfig  = require('../config/config'); //configuration of database
+
+
+const sequelize = new Sequelize(dbconfig);
+
+fs
+  .readdirSync(__dirname)
+  .filter(file => {
+    return (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js');
+  })
+  .forEach(file => {
+    // var model = sequelize['import'](path.join(__dirname, file));
+    const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
+    db[model.name] = model;
+  });
+
+Object.keys(db).forEach(modelName => {
+  if (db[modelName].associate) {
+    db[modelName].associate(db);
+  }
+});
+
+db.sequelize = sequelize;
+db.Sequelize = Sequelize;
+
+module.exports = db;
