+++ server/package.json
@@ -21,6 +21,7 @@
     "mysql": "^2.18.1",
     "mysql2": "^2.2.5",
     "nodemon": "^2.0.6",
+    "pg": "^8.5.1",
     "sequelize": "^5.22.3",
     "uuid": "^8.3.1"
   }
