+++ server/temp/index.js
@@ -1,0 +1,30 @@
+let users = {
+  1: {
+    id: '1',
+    username: 'Robin Wieruch',
+    itemIds: [1],
+  },
+  2: {
+    id: '2',
+    username: 'Dave Davids',
+    itemIds: [2],
+  },
+};
+ 
+let items = {
+  1: {
+    id: '1',
+    name: 'Hello World',
+    userId: '1',
+  },
+  2: {
+    id: '2',
+    name: 'By World',
+    userId: '2',
+  },
+};
+ 
+module.exports =  {
+  users,
+  items
+};
